Netology_django_api/rest api/smart_home/requirements.txt
django>=3.1,<4.0
djangorestframework>=3.12,<4.0
python-decouple